pyproject.toml
[project]
name = "hatena-blog-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.5.1",
    "requests>=2.32.0",
    "python-decouple>=3.8",
    "lxml>=5.0.0",
    "ruff>=0.11.11",
]
